chore(Navigation/TabletErrors): rename some files [YTFRONT-4119]
