+++ packages/ui/src/ui/pages/navigation/tabs/TabletErrors/TabletErrorsBackground.scss
@@ -1,4 +1,4 @@
-.navigation-tablet-errors {
+.navigation-tablet-errors-background {
     .elements-error:not(:last-child) {
         margin-bottom: 10px;
     }
