+++ packages/ui/src/ui/store/actions/navigation/index.js
@@ -32,7 +32,7 @@ import {getCluster, getCurrentUserName} from '../../../store/selectors/global';
 import {fetchTableMountConfig} from '../../../store/actions/navigation/content/table/table-mount-config';
 import {checkPermissions} from '../../../utils/acl/acl-api';
 import {getAnnotation} from './tabs/annotation';
-import {loadTabletErrorsCount} from './tabs/tablet-errors';
+import {loadTabletErrorsCount} from './tabs/tablet-errors-background';
 import {isSupportedEffectiveExpiration} from '../../../store/selectors/thor/support';
 import {getTabs} from '../../../store/selectors/navigation/navigation';
